Remove hardcoding of mysql adapter
Napa is somewhat mysql-centric, but it doesn't have to be. If you let the database.yml file specify `adapter: mysql`, then there is no reason to encode that into the db.rake. That will allow it to be used easily with other databases and leave the database.yml configuration to the user.